Fix compose.json for MediaWiki 1.42.4
Apparently the [a patch][1] missed the 1.42.4 which causes the
update.php script to fail:

    $ docker compose exec -it mediawiki php /var/www/html/maintenance/run.php update --quick
    MediaWiki 1.42.4 Updater
    wikimedia/parsoid: 0.19.1 installed, 0.19.0 required.
    Error: your composer.lock file is not up to date. Run "composer update --no-dev" to install newer dependencies

Fix this by manually adding the version from the patch as this is more
easily reviewable as compared to applying some random patch.

The code path is included in all docker recipies but only triggers for
the 1.42.4 release and can be removed with the next version.

[1]: https://gerrit.wikimedia.org/r/c/mediawiki/core/+/1082844

Signed-off-by: Christian Heusel <christian@heusel.eu>